fix(package): move some models to separate files, fix tests
