+++ test/unit/json_schema/specification_test.dart
@@ -37,11 +37,14 @@
 //     THE SOFTWARE.
 
 import 'dart:convert';
-import 'package:json_schema/json_schema.dart';
-import 'package:json_schema/src/json_schema/models/schema_version.dart';
-import 'package:json_schema/src/json_schema/validator.dart';
+
 import 'package:path/path.dart' as path;
 import 'package:test/test.dart';
+
+import 'package:json_schema/json_schema.dart';
+import 'package:json_schema/src/json_schema/models/schema_version.dart';
+import 'package:json_schema/src/json_schema/models/validation_results.dart';
+import 'package:json_schema/src/json_schema/validator.dart';
 
 import '../constants.dart';
 import '../specification_remotes.dart';
